refactor: view 와 관련된 컨트롤러 분리
