+++ src/main/java/roomescape/view/admin/AdminViewController.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.view.admin;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
