+++ src/main/java/roomescape/view/user/SignUpController.java
@@ -1,15 +1,16 @@
-package roomescape.member.controller.page;
+package roomescape.view.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+@RequestMapping("/signup")
 @Controller
-@RequestMapping("/login")
-public class LoginController {
+public class SignUpController {
 
     @GetMapping
-    public String getLogIn() {
-        return "login";
+    public String signUp() {
+        return "signup";
     }
+
 }
