+++ src/main/java/roomescape/view/user/UserReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.user.controller;
+package roomescape.view.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
